chore: Reorganizing the main functions.
