+++ internal/template/render/commands.go
@@ -1,0 +1,7 @@
+package render
+
+type Commands struct{}
+
+type Data struct {
+	Env map[string]any
+}
