+++ pkg/template/execute.go
@@ -1,0 +1,11 @@
+package template
+
+import "github.com/toolsascode/gomodeler/internal/template/render"
+
+func renderFile() RenderInterface {
+	return &render.Commands{}
+}
+
+func RunRender() {
+	renderFile().Run()
+}
